Login sessions are now remembered for 90 days
A cookie is set with the token and username. When going to the loging
screen, it checks if the token is valid, and starts a session if it is.

Also checks that the machine is the same, and that the token has not
expired.

Also added a logout page that erases these cookies and destroys the
session.